Create CheckBMI.cpp
The Body Mass Index (BMI) is often used by scientists and physicians to determine whether a person is underweight or overweight. The formula for calculating BMI is:

bmi  = weight  / height2
  	
Write a program that gets a person’s weight (kg) and height (m) from the keyboard and then calculates and displays the person’s BMI with two digits after the decimal. If the person’s BMI is below 20, display “You are underweight”, else if it is above 25, display “You are overweight”, else display “Your BMI is normal”.